pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spec-agent"
version = "0.2.0"
description = "Auto-generate an Obsidian knowledge wiki from every git push, powered by Claude"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Vishesh Chaitanya", email = "visheshchaitanya@gmail.com" },
]
keywords = [
    "obsidian",
    "wiki",
    "git",
    "documentation",
    "anthropic",
    "claude",
    "ai",
    "developer-tools",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Utilities",
]
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1",
    "rich>=13.0",
    "requests>=2.31",
]

[project.optional-dependencies]
ollama = [
    "requests>=2.31",
]
gemini = [
    "google-genai>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
]

[project.urls]
Homepage = "https://github.com/visheshchaitanya/spec-agent"
Repository = "https://github.com/visheshchaitanya/spec-agent"
Issues = "https://github.com/visheshchaitanya/spec-agent/issues"

[project.scripts]
spec-agent = "spec_agent.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["spec_agent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
